Link one address to parcel pseudo-identifiers for apartment complexes

To match a parcel to an apartment complex pseudo-identifier, it must contain:

a Hamilton County ZIP code

a street name matching the street names in parcel:::apt_defs

a street number within the ranges for each pseudo-identifier in parcel:::apt_defs

link_apt(x)

Arguments

x

a single address character string

Value

apt pseudo-identifier character string; NA if not matched